Use Vavr List for Matrix contents
