+++ src/main/java/com/blackcrowsys/crimson/Matrix.java
@@ -1,5 +1,6 @@
 package com.blackcrowsys.crimson;
 
+import io.vavr.collection.List;
 import lombok.Data;
 
 import java.util.ArrayList;
@@ -7,13 +8,13 @@ import java.util.ArrayList;
 @Data
 class Matrix {
 
-    private ArrayList contents;
+    private List<Double> contents;
 
     private int numOfColumns;
 
     private int numOfRows;
 
-    Matrix(ArrayList contents, int numOfColumns) {
+    Matrix(List contents, int numOfColumns) {
         this.setContents(contents);
         this.setNumOfColumns(numOfColumns);
         this.numOfRows = contents.size() / this.numOfColumns;
@@ -27,16 +28,15 @@ class Matrix {
         if (row > contents.size() / this.numOfColumns || row < 1) {
             throw new IllegalArgumentException("Incorrect Row Index: " + row);
         }
-        return (Double) contents.get(getIndex(row, column));
+        return contents.get(getIndex(row, column));
     }
 
     Matrix sum(final Matrix thatMatrix) {
         if ((this.getNumOfRows() != thatMatrix.getNumOfRows()) && (this.getNumOfColumns() != thatMatrix.getNumOfColumns()))
             throw new IllegalArgumentException("Matrix size mismatch");
         ArrayList<Double> sum = new ArrayList<>();
-        for (int i = 0; i < this.contents.size(); i++) {
-            sum.add((Double)this.contents.get(i) + (Double)thatMatrix.getContents().get(i));
-        }
+        for (int i = 0; i < this.contents.size(); i++)
+            sum.add(this.contents.get(i) + thatMatrix.getContents().get(i));
         return MatrixFactory.generateMatrix(sum, this.numOfColumns);
     }
